+++ common/src/main/java/me/pandamods/fallingtrees/api/TreeType.java
@@ -13,10 +13,16 @@
 package me.pandamods.fallingtrees.api;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 public interface TreeType {
 	boolean isTreeStem(BlockState blockState);
-	TreeData gatherTreeData(BlockPos blockPos, Level level);
+	default TreeData gatherTreeData(BlockPos blockPos, Level level) {
+		return null;
+	}
+	default TreeData gatherTreeData(BlockPos blockPos, Level level, Player player) {
+		return gatherTreeData(blockPos, level);
+	}
 }